Add mysql

+++ app/models.py
@@ -11,7 +11,7 @@ class Client(db.Model, ModelMixin, SerializerMixin):
     serialize_rules = ('-sales', '-comings', '-balances' )
     id = db.Column(db.Integer, primary_key=True)
     fullname = db.Column(db.String(100), nullable=False)
-    phone = db.Column(db.Integer, unique=True)
+    phone = db.Column(db.BigInteger, unique=True)
     productname = db.Column(db.String(100))
     status = db.Column(db.Boolean, default=True)
     comment = db.Column(db.Text)
@@ -50,8 +50,8 @@ class Sale(db.Model, ModelMixin, SerializerMixin):
     product_id = db.Column(db.Integer, db.ForeignKey('product.id', ondelete= 'CASCADE'), index=True)
     client_id = db.Column(db.Integer, db.ForeignKey('client.id', ondelete= 'CASCADE'), index=True)
     quantity = db.Column(db.Integer)
-    price = db.Column(db.Float)
-    sumprice = db.Column(db.Float)
+    price = db.Column(db.Numeric(11,2))
+    sumprice = db.Column(db.Numeric(11,2))
     sumquantity = db.Column(db.Integer)
     status = db.Column(db.Boolean, default=True)
     comment = db.Column(db.Text)
@@ -65,9 +65,9 @@ class Coming(db.Model, ModelMixin, SerializerMixin):
     date = db.Column(db.DateTime, default=datetime.datetime.now())
     product_id = db.Column(db.Integer, db.ForeignKey('product.id', ondelete= 'CASCADE'), index=True)
     quantity = db.Column(db.Integer)
-    price = db.Column(db.Float)
+    price = db.Column(db.Numeric(11,2))
     sumquantity = db.Column(db.Integer)
-    sumprice = db.Column(db.Float)
+    sumprice = db.Column(db.Numeric(11,2))
     comment = db.Column(db.Text)
     
 
@@ -81,9 +81,9 @@ class Balance(db.Model, ModelMixin, SerializerMixin):
     date = db.Column(db.DateTime, default=datetime.datetime.now())
     product_id = db.Column(db.Integer, db.ForeignKey('product.id', ondelete= 'CASCADE'), index=True)
     quantity = db.Column(db.Integer)
-    price = db.Column(db.Float)
+    price = db.Column(db.Numeric(11,2))
     sumquantity = db.Column(db.Integer)
-    sumprice = db.Column(db.Float)
+    sumprice = db.Column(db.Numeric(11,2))
     comment = db.Column(db.Text)
 
     def __repr__(self):
